% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{aic_bic_model}
\alias{aic_bic_model}
\title{Generates comprehensive output for AIC and BIC}
\usage{
aic_bic_model(final_best, sample_number)
}
\arguments{
\item{final_best}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{sample_number}{Length of the data frame/number of temperature datapoints.}
}
\description{
This function generates comprehensive output for the aic and bic statistical tests. 
It includes information such as the best model (that have the smallest aic and bic values) and the aic and bic values for each parameter model type.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
sample_number = nrow(util)
result = run_model(util)
aic_bic_values = aic_bic_model(result, sample_number)
}
