% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{aic_bic_test}
\alias{aic_bic_test}
\title{Calculates AIC and BIC}
\usage{
aic_bic_test(model, bestvalue, sample_number)
}
\arguments{
\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'.}

\item{bestvalue}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{sample_number}{Length of the data frame/number of temperature datapoints.}
}
\description{
This function generates numeric output for the aic and bic test statistical test. The ouput is a vector and the first value is AIC and the second BIC.
}
\examples{
\dontrun{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
sample_number = nrow(util)
model = '4P'
temp = sort_matrix(util$OAT,util$usage)
bestvalue = create_model(temp$x, temp$y, '5P')
aic_bic_values = aic_bic_test(model, bestvalue, sample_number)}
}
