% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit_controller.R
\name{best_worst_model}
\alias{best_worst_model}
\title{Filters model with lowest CV-RMSE}
\usage{
best_worst_model(inter_result, bdbid = NA, energy, fiscal_year = NA,
  n = NA, prepost = 1)
}
\arguments{
\item{inter_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{bdbid}{Building id, which identifies the building or the dataset. Defaults to \code{NA}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}

\item{fiscal_year}{Fiscal year. Defaults to \code{NA}.}

\item{n}{number of observed points (independent variable) in utility data frame of a building. Defaults to \code{NA}.}

\item{prepost}{A numeric value. Defaults to be 1. This indicates if it is unretrofit or retrofit (pre or post). See \code{\link{unretrofit_utility}} for more information about prepost.}
}
\description{
This function filters for model with lowest CV-RMSE value given a nested list of model result.
}
\details{
If some models pass all of shape test, t-test, and population test but do not meet threshold CV-RMSE, the model will the lowest CV-RMSE will be selected as the best model. If none of the model passes all of shape test, t-test, and population test, it will return \code{NULL}.
}
\examples{
\dontrun{
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
           & retrofit_utility$energy_type == 'Elec') #retrofit data
retrofit_result = run_model_retrofit(retrofit_util)
retrofit_best_result = best_worst_model(retrofit_result$post, bdbid = 846152, 'Elec', prepost = 3)}
}
\seealso{
\code{\link{main_best_model_func}}, \code{\link{run_model}}, and \code{\link{run_model_retrofit}}
}
