% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRema.R
\docType{package}
\name{bRema-package}
\alias{bRema}
\alias{bRema-package}
\title{bRema: Calculate and graph parameter based models for building energy usage}
\description{
Given a dataframe of outside air temperature, and energy usage (i.e utility data),
it will calculate parameters based model and graph them by using plotly.
}
\author{
\strong{Maintainer}: Tin Tun Naing \email{tinnaing347@gmail.com}

Other contributors:
\itemize{
  \item Michele Bradley \email{michele.ebradley@gmail.com} [contributor]
  \item Johnny DeBlase \email{john.deblase@cuny.edu} [contributor]
  \item Honey Berk \email{honey.berk@cuny.edu} [contributor]
  \item Krystyna Horn \email{krystyna.horn@cuny.edu} [contributor]
  \item Marco Ascazubi \email{marco.ascazubi@cuny.edu} [contributor]
  \item Daniel Ladner \email{danieljladner@gmail.com} [contributor]
  \item Kirk Shillingford [contributor]
  \item CUNY BPL [copyright holder]
}

}
\keyword{internal}
