% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{calc_cool_load}
\alias{calc_cool_load}
\title{Calculates cooling load}
\usage{
calc_cool_load(utility, cp, ycp)
}
\arguments{
\item{utility}{A utility data frame with columns, which must have these columns: energy_type, OAT, usage, end_date and day. Day column: the end date from end_date column.}

\item{cp}{cooling change-point.}

\item{ycp}{y-values at \code{cp}}
}
\description{
This function calculates cooling load.
}
