% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{calc_heat_cool_load}
\alias{calc_heat_cool_load}
\title{Calculates baseload and heating and cooling load}
\usage{
calc_heat_cool_load(percent_heating, percent_cooling, percent_baseload,
  total_consumption)
}
\arguments{
\item{percent_heating}{A numeric vector.}

\item{percent_cooling}{A numeric vector.}

\item{percent_baseload}{A numeric vector.}

\item{total_consumption}{A numeric vector.}
}
\description{
This function calculates \strong{(not predicted)} baseload and heating and cooling load by using output from \code{\link{percent_heat_cool_func}} and total_cosumption.
}
\details{
This function is different from \code{\link{predicted_heat_cool_load_func}}. See Heat and Cool Load, subsection of Methodlogy section of bRema paper.
}
