% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{calc_heat_load}
\alias{calc_heat_load}
\title{Calculates heating load}
\usage{
calc_heat_load(utility, cp, ycp)
}
\arguments{
\item{utility}{A utility data frame with columns: energy_type, OAT, usage, end_date and day. Day column: the end date from end_date column.}

\item{cp}{heating change-point.}

\item{ycp}{y-values at \code{cp}}
}
\description{
This function calculates heating load.
}
