% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{calc_loads}
\alias{calc_loads}
\title{Calculates loads}
\usage{
calc_loads(util, model_params, total_predicted_consumption,
  total_consumption, scaled = TRUE)
}
\arguments{
\item{util}{Utility data frame of a single building and single energy type, returned by \code{\link{calc_predicted_usage}}.}

\item{model_params}{List or dataframe returned by \code{\link{construct_model_table}}.}

\item{total_predicted_consumption}{Total predicted consumption (i.e sum of usages calculated using parameters from \code{model_params}).}

\item{total_consumption}{Total consumption.}

\item{scaled}{Boolean. Defaults to TRUE. Determine to multiply usage with number of days in the month.}
}
\description{
This function calculates heating load, cooling load and baseload.
}
