% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{calc_nac}
\alias{calc_nac}
\title{Calculates normalized annual comsumption}
\usage{
calc_nac(best_result, norm_temp = NULL)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{norm_temp}{A data frame of normalized average monthly temperature with columns: month and avg_temp. Default to \code{\link{norm_temp_tmy3}}.}
}
\description{
This function calculates normalized annual consumption.
}
