% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{calc_predict_heat_load}
\alias{calc_predict_heat_load}
\title{Calculates predicted heating load}
\usage{
calc_predict_heat_load(utility, cp, ycp)
}
\arguments{
\item{utility}{A utility data frame with columns: energy_type, OAT, usage, end_date, day and predicted_usage. Day column: the end date from end_date column. predicted_usage column: output from \code{\link{calc_predicted_usage}}.}

\item{cp}{heating change-point.}

\item{ycp}{y-values at \code{cp}}
}
\description{
This function calculates heating load by using predicted usage.
}
