% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{calc_predicted_usage}
\alias{calc_predicted_usage}
\title{Calculates preicted usage}
\usage{
calc_predicted_usage(util, best_df, scaled = TRUE)
}
\arguments{
\item{util}{Utility dataframe of a single building and a single energy type.}

\item{best_df}{Best Model dataframe returned by \code{\link{batch_run}}, \code{\link{batch_run_energy}} or \code{\link{construct_model_table}}.}

\item{scaled}{scaled Boolean value. Defaults to TRUE. Determine to multiply usage with number of days in the month.}
}
\description{
This function calcuates predicted usage using parameters from \code{model_params}.
}
