% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{calc_total_consumption}
\alias{calc_total_consumption}
\title{Calculates total consumption}
\usage{
calc_total_consumption(utility, scaled = TRUE)
}
\arguments{
\item{utility}{Utility dataframe returned by \code{\link{calc_predicted_usage}}.}

\item{scaled}{Boolean value. Defaults to TRUE. Determine to multiply usage with number of days in the month.}
}
\description{
This function calculates total consumption and total predicted consumption.
}
