% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{calcT}
\alias{calcT}
\title{Calculates tstat of a slope}
\usage{
calcT(slope, RMSE, xinverse)
}
\arguments{
\item{slope}{Slope of the least square line.}

\item{RMSE}{RMSE}

\item{xinverse}{The inverse matrix of independent variables.}
}
\description{
This function returns tstat given slope of the least square line, RMSE and the inverse matrix of independent variables. Only applies for model that is not '2P'.
}
\examples{
x = matrix(c(1:10))
y = matrix(c(1,2.5,3,4,5,5.5,7,8.3,9,10))
bestvalue = create_model(x, y, '4P')
slope = bestvalue$params['Slope1',1]
xsplit = splitter('4P', x, 5)
xfinal = make_matrix(xsplit)
xinverse = solve(t(xfinal)\%*\%xfinal)
RMSE = as.numeric(bestvalue$stats['RMSE',1])
tstat = calcT(slope, RMSE, xinverse) #tstat for leftmost slope of a '4P' model
}
\seealso{
\code{\link{tTest}}
}
