% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{choosing_final_best}
\alias{choosing_final_best}
\title{Chooses the best parameter model}
\usage{
choosing_final_best(best_result, weight_df)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{weight_df}{A data frame returned from \code{metric_weight} function}
}
\description{
This function returns the best model.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' 
			& unretrofit_utility$energy_type == 'Elec')
result = run_model(util)
best_result = model_pass_func(result, 'Elec')
metric_weight_df = metric_weight(best_result, TRUE)
best_model = choosing_final_best(best_result, metric_weight_df)
}
