% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{collect_slopes}
\alias{collect_slopes}
\title{Return a data frame of slopes}
\usage{
collect_slopes(best_result)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}
}
\description{
This function extracts and returns a datamframe of slopes given a nested list of individual model result.
}
