% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit_controller.R
\name{construct_all_model}
\alias{construct_all_model}
\title{Construct table for all models}
\usage{
construct_all_model(inter_result, bdbid = NA, energy, fiscal_year = NA,
  n = NA, prepost = 1)
}
\arguments{
\item{inter_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{bdbid}{Building id, which identifies the building or the dataset. Defaults to \code{NA}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}

\item{fiscal_year}{Fiscal year. Defaults to \code{NA}.}

\item{n}{number of observed points (independent variable) in utility data frame of a building. Defaults to \code{NA}.}

\item{prepost}{A numeric value. Defaults to be 1. This indicates if it is unretrofit or retrofit. See \code{\link{unretrofit_utility}} for more information about prepost.}
}
\description{
This functions returns a data frame given a nested list of model result.
}
\examples{
\dontrun{
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
           & retrofit_utility$energy_type == 'Elec') #retrofit data
retrofit_result = run_model_retrofit(retrofit_util)
pre_best_df = construct_all_model(retrofit_result$pre, bdbid = 846152, energy = 'Elec')}
}
