% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{construct_model_table}
\alias{construct_model_table}
\title{Construct model table}
\usage{
construct_model_table(best_result, bdbid = NA, energy,
  fiscal_year = NA, n = NA, prepost = 1)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{bdbid}{Building id, which identifies the building or the dataset. Defaults to \code{NA}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}

\item{fiscal_year}{Fiscal year. Defaults to \code{NA}.}

\item{n}{number of observed points (independent variable) in utility data frame of a building. Defaults to \code{NA}.}

\item{prepost}{A numeric value. Defaults to be 1. This indicates if it is unretrofit or retrofit. See \code{\link{unretrofit_utility}} for more information about prepost.}
}
\description{
This functions returns a data frame given a nested list of individual model result.
}
\section{Note}{

If more than half of energy usage points of a facility are zeros, that faciltiy will be skipped.
}

\examples{

util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
inter_result = run_model(util)
best_result = main_best_model_func(inter_result)
construct_model_table(best_result, 'f3acce86', 'Elec', 2017)
construct_model_table(inter_result[['3PC']], 'f3acce86', 'Elec', 2017)
}
