% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{create_model}
\alias{create_model}
\title{Calculates the best segmented regression parameters for a given model}
\usage{
create_model(x, y, model, step = 0.5)
}
\arguments{
\item{x}{A (column) matrix of outside air temperature, independent variables.}

\item{y}{A (column) matrix of actual values(usage), dependent variables.}

\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'.}

\item{step}{A single numeric value. Defaults to 0.5.}
}
\value{
A list with components:
\item{stats}{A matrix of statistical results such as RMSE, R-squared}
\item{params}{A matrix of parameters such as basload and slopes}
\item{cp1}{Change point. The leftmost change point. This will be zero if \code{model} is '2P'.}
\item{cp2}{Change point. The rightmost change point. This will be zero if \code{model} is not '5P'.}
\item{model}{Model type.}
}
\description{
This function returns the best change-point parameter based model given a specific model and independent and dependent column matrices by using only one step size.
This function is dependent on \code{test_model} and \code{splitter} function.
}
\examples{
x = matrix(c(1:10))
y = matrix(c(1,2.5,3,4,5,5.5,7,8.3,9,10))
result = create_model(x, y, '4P')
}
\seealso{
\code{\link{create_model_2}}, and \code{\link{run_model}}
}
