% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{heat_cool_sen_change}
\alias{heat_cool_sen_change}
\title{Extract heating/cooling sensitivity and change-points}
\usage{
heat_cool_sen_change(best_df, model, energy)
}
\arguments{
\item{best_df}{Best model data frame (of a single builiding) returned from \code{\link{batch_run}}.}

\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'. The model must be present in \code{best_df}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}
}
\description{
This function extracts slopes and change-points from \code{best_model} data frame and renames them meaningfully depending on model. For example, if it is '3PC' model, right slope (rs) becomes cooling senitivity.
}
\examples{
\dontrun{
best_df_'f3acce86' = subset(best_df, best_df$bdbid == 'f3acce86' & best_df$energy_type == 'Elec')
model = as.character(best_df_'f3acce86'$model_type)
heat_cool_sen_change(best_df_'f3acce86', model, 'Elec')
#compare the result with best_df_'f3acce86' to see what have been renamed.
}
}
