% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{lean_analysis_energy}
\alias{lean_analysis_energy}
\title{Post model and Lean Analysis for specific energy type}
\usage{
lean_analysis_energy(utility, best_model, energy, lean_category_n = NA,
  fiscal_year_n = NA, rank_flag = TRUE)
}
\arguments{
\item{utility}{A data frame of \strong{unretrofit data} with multiple buildings, with colums: OAT, usage, energy_type and end_date. See \code{\link{unretrofit_utility}} for data format.}

\item{best_model}{A data frame returns from \code{\link{batch_run}} or \code{\link{batch_run_energy}}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}

\item{lean_category_n}{A character string or a numeric value to rank buildings in groups. Defaults to \code{NA}.}

\item{fiscal_year_n}{A numeric value. Defaults to \code{NA}.}

\item{rank_flag}{A boolean value. Defaults to \code{TRUE}. If set to \code{FALSE}, there will be no ranking of baseload, heating/cooling change-points and sensitivity.}
}
\description{
This function transforms best_model in a more meaningful way and returns lean analysis calculation results after modeller, such as numeric and percent ranks for heating/cooling change-points, heating/cooling sensitivity and baseload.
}
\examples{
\dontrun{
batch_result = batch_run(unretrofit_utility)
post_df_elec = lean_analysis_energy(unretrofit_utility, batch_result$best_result_df, 'Elec')}
}
