% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{lean_analysis_ranking}
\alias{lean_analysis_ranking}
\title{Ranks baseload, heating/cooling change-points and sensitivity.}
\usage{
lean_analysis_ranking(post_df)
}
\arguments{
\item{post_df}{A data frame returned from \code{\link{main_post_model}}. The data frame should include multiple buildings to have meaningful interpretation of the rank.}
}
\description{
This function calculates numeric and percent rank for baseload, heating/cooling change-points and sensitivity (for a sinlge energy type).
}
\details{
If \code{post_df} includes more than one energy type, it is advised to use \code{\link{main_post_model}}.
}
\examples{
\dontrun{
batch_result = batch_run(unretrofit_utility)
post_df = main_post_model(unretrofit_utility, batch_result$best_result_df)
post_elec = subset(post_df, post_df$energy_type == 'Elec')
lean_elec = lean_analysis_ranking(post_elec)
}
}
