% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{least_squares}
\alias{least_squares}
\title{Linear regression}
\usage{
least_squares(xmat, ymat)
}
\arguments{
\item{xmat}{Independent matrix (dimesions, either n by 2 or n by 3), where n is number of sample points. Matrix returned by \code{make_matrix} function.}

\item{ymat}{Dependent matrix (dimesions, n by 1), where n is number of sample points.}
}
\value{
A matrix with components:
\item{ycp}{y-value at the change-point}
\item{Slope1}{leftmost slope if B is 3 by 1 matrix. Otherwise, non-zero slope.}
\item{Slope2}{rightmost slope if if B is 3 by 1 matrix.}
}
\description{
This function returns parameters of least square equation given column matrices of independent variables and dependent matrices.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
temp = sort_matrix(util$OAT,util$usage)
x_split = splitter('4P', temp$x, 51)
xmat = make_matrix(x_split)
B = least_squares(xmat, temp$y)
}
\seealso{
\code{\link{make_matrix}}
}
