% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{main_best_model_func}
\alias{main_best_model_func}
\title{Returns the best model}
\usage{
main_best_model_func(inter_result, energy_n, metric_flag = TRUE,
  CV_RMSE_elec = 0.25, Rsquared_elec = 0.75, CV_RMSE_fuel = 0.5,
  Rsquared_fuel = 0.75)
}
\arguments{
\item{inter_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{energy_n}{A charcter string. Energy Type, either 'Elec' or 'Fuel'.}

\item{metric_flag}{A boolean value. Defaults to TRUE. If set to \code{TRUE}, \code{\link{order_models}} function will be use for scoring models for 'Elec' and \code{\link{inverse_order_models}} for 'Fuel'. Else, \code{inverse_order_models} function will be used for 'Elec' and \code{order_models} for 'Fuel'.}

\item{CV_RMSE_elec}{A numeric value. CV-RMSE threshold value for Elec. Defaults to 0.25.}

\item{Rsquared_elec}{A numeric value. Rsquared threshold value for Elec. Defaults to 0.75.}

\item{CV_RMSE_fuel}{A numeric value. CV-RMSE threshold value for Fuel. Defaults to 0.50.}

\item{Rsquared_fuel}{A numeric value. Rsquared threshold value for Fuel. Defaults to 0.75.}
}
\value{
A list with components:
\item{stats}{A matrix of statistical results such as RMSE, R-squared}
\item{params}{A matrix of parameters such as basload and slopes}
\item{cp1}{Change point. The leftmost change point. This will be zero if \code{model} is '2P'.}
\item{cp2}{Change point. The rightmost change point. This will be zero if \code{model} is not '5P'.}
\item{model}{Model type.}
\item{stat_test$main_test}{Main(final) test result. Will be 'Fail' if the model fails one of tTest, population test or shape test. Else, 'Passs'.}
\item{stat_test$tTest}{t-test results}
\item{stat_test$pop_test}{population test results}
\item{stat_test$shape_test}{shape test results}
\item{figure}{A parameter graph.}
}
\description{
This function returns the best model if there is one. Otherwise, return an empty list.
}
\examples{
\dontrun{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec') #unretrofit data
result = run_model(util)
best_result = main_best_model_func(result, 'Elec') #best_model for unretrofit data
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
					& retrofit_utility$energy_type == 'Elec') #retrofit data
retrofit_result = run_model_retrofit(retrofit_util)
retrofit_result$pre #result for pre-retrofit

#best_model for pre-retrofit data
retrofit_best_result = main_best_model_func(retrofit_result$pre, 'Elec')}
}
\seealso{
\code{\link{run_model}}, and \code{\link{run_model_retrofit}}
}
