% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit_controller.R
\name{main_best_model_retrofit}
\alias{main_best_model_retrofit}
\title{Returns the best model for retrofit data}
\usage{
main_best_model_retrofit(inter_result, energy, metric_vec = c(1, 2, 3, 4, 5),
  CV_RMSE_n = 0.25, Rsquared_n = 0.75)
}
\arguments{
\item{inter_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{energy}{A charcter string. Energy Type, either 'Elec' or 'Fuel'.}

\item{metric_vec}{A numeric (integer) vector. This vector determines the pirority of model types. Defaults to \code{c(1,2,3,4,5)} for models: '2P', '3PC', '3PH', '4P', '5P'.}

\item{CV_RMSE_n}{A numeric value. CV-RMSE threshold value. Defaults to 0.25.}

\item{Rsquared_n}{A numeric value. Rsquared threshold value. Defaults to 0.75.}
}
\description{
This function returns the best model if there is one for retrofit data. Otherwise, return a null list.
}
\details{
If some models pass all of shape test, t-test, and population test but do not meet threshold CV-RMSE, the model will the lowest CV-RMSE will be selected as the best model. If none of the model passes all of shape test, t-test, and population test, it will return a list with null component.
}
\examples{
\dontrun{
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
           & retrofit_utility$energy_type == 'Elec') #retrofit data
retrofit_result = run_model_retrofit(retrofit_util)
retrofit_best_result = main_best_model_retrofit(retrofit_result, 'Elec')}
}
\seealso{
\code{\link{main_best_model_func}}, \code{\link{run_model}}, and \code{\link{run_model_retrofit}}
}
