% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{main_heat_cool_load}
\alias{main_heat_cool_load}
\title{Calculates heating and cooling load}
\usage{
main_heat_cool_load(utility, cp1 = 0, cp2 = 0, ycp, model)
}
\arguments{
\item{utility}{A utility data frame of \strong{unretrofit data} with columns: energy_type, OAT, usage, end_date and day. Day column: the end date from end_date column.}

\item{cp1}{A numeric value. The first change-point. Defaults to 0.}

\item{cp2}{A numeric value. The second change-point. Defaults to 0.}

\item{ycp}{y-value at \code{cp1} and \code{cp2}.}

\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'.}
}
\description{
This function calculates heating and cooling load of a builiding.
}
