% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{main_lean_analysis}
\alias{main_lean_analysis}
\title{Main logic for lean analysis}
\usage{
main_lean_analysis(post_df)
}
\arguments{
\item{post_df}{A data frame returned from \code{\link{main_post_model}}. The data frame should include multiple buildings to have meaningful interpretation of the rank.}
}
\description{
Main logic for lean analysis. Can handle multiple energy types.
}
\examples{
\dontrun{
batch_result = batch_run(unretrofit_utility)
post_df = main_post_model(unretrofit_utility, batch_result$best_result_df)
lean_df = main_lean_analysis(post_df)
}
}
