% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{main_numeric_percent_rank}
\alias{main_numeric_percent_rank}
\title{Ranks baseload, heating/cooling change-points and sensitivity.}
\usage{
main_numeric_percent_rank(post_df, energy)
}
\arguments{
\item{post_df}{A data frame returned from \code{\link{post_model_prelimanary}}. The data frame should include multiple buildings to have meaningful interpretation of the rank.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}
}
\description{
This function calculates numeric and percent rank for baseload, heating/cooling change-points and sensitivity.
}
