% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{main_saving_calc}
\alias{main_saving_calc}
\title{Handles calculation for savings and savings uncertainty}
\usage{
main_saving_calc(utility, best_model, session_id = NA)
}
\arguments{
\item{utility}{A utility data frame of \strong{retrofit data} with multiple buildings.}

\item{best_model}{best_model A data frame returns from \code{\link{batch_run}} or \code{\link{batch_run_energy}}}

\item{session_id}{Either a character string or a numeric value to keep track. Defaults to \code{NA}.}
}
\value{
A list with components:
\item{saving_df}{A data frame of saving calculation}
\item{uncertainty_df}{A data frame of saving uncertainty calculation}
}
\description{
This function handles calculation for savings and savings uncertainty for \strong{retrofit data}.
}
\examples{
\dontrun{
batch = batch_run_retrofit(retrofit_utility)
saving_list = main_saving_calc(retrofit_utility, batch$best_result_df)
}
}
