% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{make_matrix}
\alias{make_matrix}
\title{Restructures a matrix returned by \code{splitter} function}
\usage{
make_matrix(xsplit)
}
\arguments{
\item{xsplit}{A column matrix (manipulated by \code{splitter} function based on change-point(s) and model).}
}
\description{
This function adds a column of zeros and ones to the matrix returned by \code{splitter} function.
The output is used in \code{least_squares} function as independent matrix input.
}
\examples{
x = matrix(c(1,2.3,2.78, 3.4,5.5, 6.1, 7 ,8.4, 9,10))
x_split = splitter('4P', x, 3)
x_final = make_matrix(x_split)
}
\seealso{
\code{\link{splitter}}
}
