% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{make_post_model_table}
\alias{make_post_model_table}
\title{Consturct post model table}
\usage{
make_post_model_table(utility, model_params, scaled = TRUE)
}
\arguments{
\item{utility}{Utility dataframe. See \code{\link{unretrofit_utility}}.}

\item{model_params}{Best Model dataframe returned by \code{\link{batch_run}}, \code{\link{batch_run_energy}} or \code{\link{construct_model_table}}.}

\item{scaled}{Boolean value. Defaults to TRUE. Determine to multiply usage with number of days in the month.}
}
\description{
This function makes post model table for a single building and a single energy type. For multiple buildings with multiple energy types, use \code{\link{main_post_model}}.
}
