% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{matrix_length}
\alias{matrix_length}
\title{Checks the length of a matrix}
\usage{
matrix_length(xsplit)
}
\arguments{
\item{xsplit}{A column matrix (manipulated by \code{splitter} function based on change-point(s) and model).}
}
\description{
This function checks if less than 25 percent of the entries of a matrix is zero or not. It returns \code{FALSE} only if less than 25 percent of matrix entries is non-zero.
}
\examples{
x = matrix(c(1,2.3,2.78, 3.4,5.5, 6.1, 7 ,8.4, 9,10))
x_split = splitter('4P', x, 3)
matrix_length(x_split)
}
