% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{metric_weight}
\alias{metric_weight}
\title{Scores models based on energy and model types}
\usage{
metric_weight(best_result, flag = TRUE)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{flag}{A boolean value. Defaults to \code{TRUE}. If set to \code{TRUE}, \code{\link{order_models}} will be used. Else, \code{\link{inverse_order_models}} function will be used. It is recommened that to set this flag to \code{TRUE} for 'Elec' and \code{FALSE} for 'Fuel'}
}
\description{
This function returns a data frame in which best models are given a metric weight depending on energy type and model type.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
result = run_model(util)
best_result = model_pass_func(result, 'Elec')
metric_weight_df = metric_weight(best_result, TRUE)
}
\seealso{
\code{\link{main_best_model_func}}
}
