% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{model_pass_func}
\alias{model_pass_func}
\title{Filters for models that have satisfying threshold}
\usage{
model_pass_func(inter_result, CV_RMSE_n = 0.25, Rsquared_n = 0.75)
}
\arguments{
\item{inter_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{CV_RMSE_n}{A numeric value. CV-RMSE threshold value. Defaults to 0.25.}

\item{Rsquared_n}{A numeric value. Rsquared threshold value. Defaults to 0.75.}
}
\description{
This function returns a list which contains models that have satisfying threshold. Defualts threshold: CV-RMSE <= 0.25 and r-squared >= 0.75, and Pass all of shape test, tTest and population test. Suggested Threshold value for fuel is CV-RMSE <= 0.50 and r-squared >= 0.75; for elec, one can use default value.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
result = run_model(util)
best_result = model_pass_func(result)
}
