% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{model_test}
\alias{model_test}
\title{Tests if a model passes tTest, population test, and shape test}
\usage{
model_test(x, y, bestvalue, model, n = 4)
}
\arguments{
\item{x}{A vector. Independent variables.}

\item{y}{A vector. Dependent variables.}

\item{bestvalue}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'.}

\item{n}{A numeric value that determines threshold for population test: \code{thereshold = number_of_independent_variables/n}. Defaults to 4. See \code{\link{pop_test}}.}
}
\description{
This function determines if a given model passes tTest, population test, and shape test. The output is a nested list. If the model passes all three tests, 'Pass' is returned as \code{main_test} result; otherwise, as 'Fail'.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
temp = sort_matrix(util$OAT,util$usage)
bestvalue = create_model(temp$x, temp$y, '5P')
test_result = model_test(temp$x, temp$y, bestvalue, '5P')
}
