% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norm_temp_tmy3}
\alias{norm_temp_tmy3}
\title{Normalized average monthly temperature}
\format{A data frame with 12 rows and 2 variables:
\describe{
\item{month}{month}
\item{avg_temp}{normalized average monthly temperature in Fahrenheit}
}}
\usage{
norm_temp_tmy3
}
\description{
A dataset containing normalized average monthly temperature (unit F) TMY3 from Laguardia airport.
}
\keyword{datasets}
