% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{order_models}
\alias{order_models}
\title{Scores models for Elec}
\usage{
order_models(best_result, i, df)
}
\arguments{
\item{best_result}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{i}{A numeric value.}

\item{df}{A data frame.}
}
\description{
This function returns a metric weight score for Elec according to model type. This function is used in \code{metric_weight} function.
}
\seealso{
\code{\link{metric_weight}}
}
