% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{percent_heat_cool_func}
\alias{percent_heat_cool_func}
\title{Calculates percentanges of heat/cool load and baseload}
\usage{
percent_heat_cool_func(predicted_total_consumption, predicted_heat_load,
  predicted_cool_load)
}
\arguments{
\item{predicted_total_consumption}{total predicted consumption. To calculate predicted_total_consumption, use output returned from \code{\link{calc_predicted_usage}}.}

\item{predicted_heat_load}{predicted heating load. Use output return from \code{\link{calc_predict_heat_load}} or \code{\link{predict_heat_cool_load_func}}.}

\item{predicted_cool_load}{preidcted cooling load. Use output return from \code{\link{calc_predict_cool_load}} or \code{\link{predict_heat_cool_load_func}}.}
}
\description{
This function calculates percentanges of heat/cool load and baseload by using output from \code{\link{predict_heat_cool_load_func}}.
}
\examples{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df
post_df = post_model_preliminary(util, best_df, bdbid_n, energy)
percent_df = percent_heat_cool_func(post_df$predicted_total_consumption,
                                     post_df$predicted_heat_load, post_df$predicted_cool_load)
}
