% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_point}
\alias{plot_point}
\title{Handles plotting points for parameter models}
\usage{
plot_point(df, energy, pre_key = 0, model_fig = plot_ly(),
  b_name = "")
}
\arguments{
\item{df}{A data frame with columns: x, y, z (OAT, usage, estimated from utillity data frame)}

\item{energy}{A character string. Energy Type, 'Elec' or 'Fuel'.}

\item{pre_key}{A numeric value. 0 for unretrofit, 1 for pre-retrofit and 3 for post-retrofit. Defaults to 0.}

\item{model_fig}{A plotly object. Defaults to \code{plot_ly()}.}

\item{b_name}{Plot title (Building Name or Buidling ID). Defaults to an empty string.}
}
\description{
This function returns a scatter plot.
}
\seealso{
\code{\link{plot_model}}
}
