% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{post_model_preliminary}
\alias{post_model_preliminary}
\title{Performs preliminary post model calculation and rearranging}
\usage{
post_model_preliminary(util, best_df, bdbid_n = NA, energy)
}
\arguments{
\item{util}{A data frame with columns: OAT, usage, and end_date. See \code{\link{unretrofit_utility}} for more information about data format.}

\item{best_df}{Best model data frame (of a single builiding) returned from \code{\link{batch_run}}.}

\item{bdbid_n}{Building id, which identifies the building or the dataset. Defaults to \code{NA}.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}
}
\description{
This function transforms \code{best_model} in a more meaningful way and returns calculations such as heating load and cooling load.
}
\examples{
\dontrun{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df
post_df = post_model_preliminary(util, best_df, bdbid_n, energy)
}
}
