% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{predict_heat_cool_load_2P}
\alias{predict_heat_cool_load_2P}
\title{Calculates heating and cooling load for 2P}
\usage{
predict_heat_cool_load_2P(util, predicted_total_consumption, slope)
}
\arguments{
\item{util}{A utility data frame of \strong{unretrofit data} with columns: energy_type, OAT, usage, end_date, day and predicted_usage. Day column: the end date from end_date column.}

\item{predicted_total_consumption}{A numeric value. Gross total predicted usage.}

\item{slope}{A numeric value.}
}
\description{
This function calculates predicted heating and cooling load of a builiding for '2P', model type.
}
\details{
Since 2P does not have any changepoints, heating load and cooling load are calculated differently than other models. See Heat and Cool Load, subsection of Methodlogy section of bRema paper. For other model types, see \code{\link{predict_heat_cool_load_func}}.
}
\examples{
\dontrun{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df 
slope = subset(best_df$ls, best_df$bdbid == 'f3acce86' &
         best_df$energy_type == 'Elec')
util$predicted_usage = calc_predicted_usage(util, best_df)
predicted_total_consumption = sum(util$predicted_usage * util$day)
predicted_load_df = predict_heat_cool_load_2P(util,
           predicted_total_consumption, slope)
}
}
