% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{predict_heat_cool_load_func}
\alias{predict_heat_cool_load_func}
\title{Calculates heating and cooling load}
\usage{
predict_heat_cool_load_func(utility, model, ycp, cp1 = NA, cp2 = NA)
}
\arguments{
\item{utility}{A utility data frame of \strong{unretrofit data} with columns: energy_type, OAT, usage, end_date, day and predicted_usage. Day column: the end date from end_date column.}

\item{model}{A character string. Model: '3PH', '3PC', '4P' or '5P'.}

\item{ycp}{y-value at \code{cp1} and \code{cp2}.}

\item{cp1}{A numeric value. The first change-point. Defaults to NA.}

\item{cp2}{A numeric value. The second change-point. Defaults to NA.}
}
\description{
This function calculates predicted heating and cooling load of a builiding using predicted usage.
}
\details{
For '2P' model, see \code{\link{predict_heat_cool_load_2P}}.
}
\examples{
\dontrun{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df 
util$predicted_usage = calc_predicted_usage(util, best_df)
predicted_load_df = predict_heat_cool_load_func(util, best_df$model,
                             best_df$ycp, best_df$cp1, best_df$cp2)
}
}
