% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{run_model}
\alias{run_model}
\title{Calculates and graphs parameter based models}
\usage{
run_model(util, plot_flag = FALSE, step = 0.5, n = 4,
  unretrofit_flag = TRUE)
}
\arguments{
\item{util}{A data frame with columns columns: energy_type, OAT, usage and estimated (1 for estimated usage and 0 for actual usage). See \code{\link{unretrofit_utility}} for more information about data format.}

\item{plot_flag}{A boolean value. Defaults to \code{FALSE}. If set to \code{FALSE}, it will not return any plots.}

\item{step}{A single numeric value or a vector. If it is a single numeric value, \code{create_model} will be used to find change-point(s). If it is a vector with more than one component, \code{create_model_2} will be used. Defaults to 0.5. Adjust \code{step}-size depending on independent varaiables.}

\item{n}{A numeric value that determines threshold for population test: \code{thereshold = number_of_independent_variables/n}. Defaults to 4. See \code{\link{pop_test}}.}

\item{unretrofit_flag}{A boolean value to indicate whether it is retrofit or unretrofit. Defaults to \code{TRUE}. This only affects parameter model graph's legend labels.}
}
\description{
This function returns parameters, stats, change-point(s), and parameter model graphs using \code{create_model} function.
}
\details{
This function can be used for both retrofit and unretrofit data.
However, if you want to run this function with retrofit data, \code{subset} or \code{filter} the data first by prepost.
If input \code{util} data frame is not in the exact format as \code{\link{unretrofit_utility}}, change independent varaible column to 'OAT' and dependent variable to 'usage', add 'energy_type' column and set all the values to either 'Fuel' or 'Elec'. If 'estimated' is column is not present, a column of ones will be added automatically.
}
\examples{
\dontrun{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec') #unretrofit data
result = run_model(util) #result for unretrofit data
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
					& retrofit_utility$energy_type == 'Elec') #retrofit data
pre_util = subset(retrofit_util, retrofit_util$prepost == 1)#subset pre-retrofit utility data
pre_result = run_model(pre_util) #result for pre-retrofit data
post_util = subset(retrofit_util, retrofit_util$prepost == 3)#subset post-retrofit utility data
post_result = run_model(post_util) #result for post-retrofit data}
}
\seealso{
\code{\link{create_model}}, \code{\link{create_model_2}}, \code{\link{run_model_retrofit}}, \code{\link{unretrofit_utility}} and \code{\link{retrofit_utility}}
}
