% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrofit_controller.R
\name{run_model_retrofit}
\alias{run_model_retrofit}
\title{Calculates and graphs parameter based models for retrofit data}
\usage{
run_model_retrofit(util, plot_flag = FALSE, step = 0.5, n_pre = 4,
  n_post = 4)
}
\arguments{
\item{util}{A data frame with columns: energy_type, OAT, usage, prepost (1 for pre-retrotit, 2 for retrofit and 3 for post-retrofit) and estimated (1 for estimated usage and 0 for actual usage). See \code{\link{retrofit_utility}} for more information about data format.}

\item{plot_flag}{A boolean value. Defaults to \code{FALSE}. If set to \code{FALSE}, it will not return any plots.}

\item{step}{A single numeric value or a vector. If it is a single numeric value, \code{create_model} will be used to find change-point(s). If it is a vector with more than one component, \code{create_model_2} will be used. Defaults to 0.5. Adjust \code{step}-size depending on independent varaiables.}

\item{n_pre}{A numeric value that determines threshold for population test for pre-retrofit: \code{thereshold = number_of_independent_variables/n}. Defaults to 4. See \code{\link{pop_test}}.}

\item{n_post}{A numeric value that determines threshold for population test for post-retrofit: \code{thereshold = number_of_independent_variables/n}. Defaults to 4. See \code{\link{pop_test}}.}
}
\description{
This function returns parameters, stats, change-point(s), and parameter model graphs using \code{run_model} function. Do not use this function for unretrofit data.
If input \code{util} data frame is not in the exact format as \code{\link{unretrofit_utility}}, change independent varaible column to 'OAT' and dependent variable to 'usage', add 'energy_type' column and set all the values to either 'Fuel' or 'Elec'. If 'estimated' is column is not present, a column of ones will be added automatically.
}
\examples{
\dontrun{
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
           & retrofit_utility$energy_type == 'Elec') #retrofit data
retrofit_result = run_model_retrofit(retrofit_util)}
}
\seealso{
\code{\link{run_model}} and \code{\link{retrofit_utility}}
}
