% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{saving_uncertainty_calc}
\alias{saving_uncertainty_calc}
\title{Calculates saving uncertainity}
\usage{
saving_uncertainty_calc(num_pre, num_post, rmse, yearly_savings, energy,
  bdbid = NA, session_id = NA)
}
\arguments{
\item{num_pre}{A numeric value. Number of pre retrofit data points.}

\item{num_post}{A numeric value. Number of post retrofit data points.}

\item{rmse}{A numeric value. Pre-retrofit RMSE-value.}

\item{yearly_savings}{A numeric value. Yearly savings returned from \code{\link{saving_calc}}.}

\item{energy}{A charcter string. Energy Type, either 'Elec' or 'Fuel'.}

\item{bdbid}{Building id, which identifies the building or the dataset. Defaults to \code{NA}.}

\item{session_id}{Either a character string or a numeric value to keep track. Defaults to \code{NA}.}
}
\description{
This function calculates saving uncertainty for \strong{retrofit data}.
}
\examples{
\dontrun{
retrofit_util = subset(retrofit_utility, retrofit_utility$bdbid == 846152
				& retrofit_utility$energy_type == 'Elec')
result = batch_run_energy_retrofit(retrofit_util, 'Elec')

#subsetting for pre-retrofit period
pre_result = subset(result$best_result_df, result$best_result_df$prepost == 1)
pre_retrofit_util = subset(retrofit_util, retrofit_util$prepost == 3)
savings_df = saving_calc(pre_retrofit_util, pre_result, 846152, 'Elec')

rmse = .subset2(pre_result, 'rmse')[1]
yearly_savings = .subset2(savings_df, 'savings')[4]
num_pre = nrow(subset(retrofit_util, retrofit_util$prepost == 1))
num_post = nrow(subset(retrofit_util, retrofit_util$prepost == 3))
uncertainty_df = saving_uncertainty_calc(num_pre, num_post, rmse, yearly_savings,
										'Elec', 846152)
}
}
