% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sorts matrices}
\usage{
sort_matrix(x, y = matrix(0))
}
\arguments{
\item{x}{A vector of independent variables.}

\item{y}{A vector of dependent variables. Defaults to zero matrix.}
}
\description{
This function sorts vector \code{x} in increasing order and sorts vector \code{y} in the order of sorted \code{x}.
The output is a list with two matrices, \code{x} and \code{y}.
}
\examples{
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86'
			& unretrofit_utility$energy_type == 'Elec')
temp = sort_matrix(util$OAT,util$usage)
xmat = temp$x
ymat = temp$y
}
