% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{stats}
\alias{stats}
\title{Calculates statistics}
\usage{
stats(y, y_predict)
}
\arguments{
\item{y}{A (column) matrix of actual values(usage).}

\item{y_predict}{A (column) matrix of predicted values(usage).}
}
\description{
This function returns a matrix of 'RMSE', 'Rsquared', 'CV_RMSE', 'SSE', 'SST', 'MSE' given actual and predicted values.
}
\examples{
y_predict = matrix(c(1:10))
y = matrix(c(1,2.5,3,4,5,5.5,7,8.3,9,10))
stat_matrix = stats(y, y_predict)
}
