% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller.R
\name{tTest}
\alias{tTest}
\title{tTest}
\usage{
tTest(x, bestvalue, model)
}
\arguments{
\item{x}{A vector. Independent variables.}

\item{bestvalue}{A list containing information about parameters such as slopes, change-points, and stats such as RMSE.}

\item{model}{A character string. Model such as '2P', '3PH', '3PC', '4P' or '5P'.}
}
\description{
This function returns a matrix of "Pass" or "Fail" depending on tstat value(s) returned by \code{calcT} function. This function is called inside \code{model_test}. The threshold for tstat is set to 2 standard deviation from the null hypothesis where the slope is equal to zero.
}
\examples{
x = matrix(c(1:10))
y = matrix(c(1,2.5,3,4,5,5.5,7,8.3,9,10))
model = '4P'
bestvalue = create_model(x, y, model)
t_matrix = tTest(x, bestvalue, model)
}
\seealso{
\code{\link{calcT}}, \code{\link{pop_test}} and \code{\link{shape_test}}
}
