% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRema.R
\docType{package}
\name{bRema-package}
\alias{bRema}
\alias{bRema-package}
\title{bRema: Building Energy Modelling and Analysis Tool in R}
\description{
Given a dataframe of outside air temperature, and energy usage (i.e utility data),
it will calculate parameters based model and graph them by using plotly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cunybpl/brema}
  \item Report bugs at \url{https://github.com/cunybpl/brema/issues}
}

}
\author{
\strong{Maintainer}: Tin Tun Naing \email{tinnaing347@gmail.com}

Other contributors:
\itemize{
  \item Michele Bradley \email{michele.ebradley@gmail.com} [contributor]
  \item Johnny DeBlase \email{jdeblase@ccny.cuny.edu} [contributor]
  \item Honey Berk \email{hberk@ccny.cuny.edu} [contributor]
  \item Krystyna Horn \email{khorn@ccny.cuny.edu} [contributor]
  \item Marco Ascazubi \email{mascazubi@ccny.cuny.edu} [contributor]
  \item Daniel Ladner \email{danieljladner@gmail.com} [contributor]
  \item Kirk Shillingford [contributor]
  \item CUNY BPL [copyright holder]
}

}
\keyword{internal}
