% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{calc_predicted_usage}
\alias{calc_predicted_usage}
\title{Calculates percentanges of heat/cool load and baseload}
\usage{
calc_predicted_usage(util, best_df)
}
\arguments{
\item{util}{A data frame with columns: OAT, usage, and end_date. See \code{\link{unretrofit_utility}} for more information about data format.}

\item{best_df}{Best model data frame (of a single builiding) returned from \code{\link{batch_run}}.}
}
\description{
This function calculates predicted usage (of a single building) based on modeller result. The output \code{predicted_usage} is used in \code{\link{predict_heat_cool_load_func}}
}
\examples{
\dontrun{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df
util$predicted_usage = calc_predicted_usage(util, best_df)
}
}
