% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{lean_analysis_ranking}
\alias{lean_analysis_ranking}
\title{Ranks baseload, heating/cooling change-points and sensitivity.}
\usage{
lean_analysis_ranking(post_df, energy)
}
\arguments{
\item{post_df}{A data frame returned from \code{\link{post_model_prelimanary}}. The data frame should include multiple buildings to have meaningful interpretation of the rank.}

\item{energy}{A character string. Energy Type, either 'Elec' or 'Fuel'.}
}
\description{
This function calculates numeric and percent rank for baseload, heating/cooling change-points and sensitivity.
}
\examples{
\dontrun{
batch_result = batch_run(unretrofit_utility)
post_df = main_post_model(unretrofit_utility, batch_result$best_result_df)
lean_elec = lean_analysis_ranking(unretrofit_utility, 'Elec')
}
}
