% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_calc.R
\name{main_post_model}
\alias{main_post_model}
\title{main_post_model}
\usage{
main_post_model(utility, best_model, scaled = TRUE)
}
\arguments{
\item{utility}{Utility dataframe. See \code{\link{unretrofit_utility}} for the format.}

\item{best_model}{Best Model dataframe returned by \code{\link{batch_run}}, \code{\link{batch_run_energy}} or \code{\link{construct_model_table}}.}

\item{scaled}{Boolean value. Defaults to TRUE. Determine to multiply usage with number of days in the month.}
}
\description{
Main function for generating post model output;  transforms best_model in a more meaningful way and calculates loads. Also can be used for retrofit data.
}
\examples{
\dontrun{
batch_result = batch_run(unretrofit_utility)
post_df = main_post_model(unretrofit_utility, batch_result$best_result_df)
}
}
