% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model_and_lean_calc.R
\name{percent_heat_cool_func}
\alias{percent_heat_cool_func}
\title{Calculates percentanges of heat/cool load and baseload}
\usage{
percent_heat_cool_func(consumption, heat_load, cool_load)
}
\arguments{
\item{consumption}{total consumption.}

\item{heat_load}{heating load.}

\item{cool_load}{cooling load.}
}
\description{
This function calculates baseload and percentanges of heat/cool load and baseload.
}
\examples{
\dontrun{
bdbid_n = 'f3acce86'
energy = 'Elec'
#first subsetting utility data frame by bdbid and energy
util = subset(unretrofit_utility, unretrofit_utility$bdbid == 'f3acce86' &
               unretrofit_utility$energy_type == 'Elec')
#finding best model and calculating best model data frame
best_df = batch_run_energy(util, 'Elec')$best_result_df
post_df = post_model_prelimanary(util, best_df, bdbid_n, energy)
percent_heat_cool_func(post_df$total_consumption, post_df$heat_load, post_df$cool_load)
}
}
